% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_survival.R
\name{cond_survival}
\alias{cond_survival}
\title{Conditional survival matrix and O_tau}
\description{
TAUS (Target-Age Unified Survival): conditional survival P(T>tau|T>t) and
the probability O_tau that a randomly sampled individual of unknown age
will outlive tau.
}
\keyword{internal}
