% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{f_t_census}
\alias{f_t_census}
\title{Histogram of times to event, coloured by exit reason and faceted by a variable of interest}
\usage{
f_t_census(
  data,
  time_var,
  event_var,
  var,
  nbins = max(data[[time_var]]),
  ncol = 1
)
}
\arguments{
\item{data}{A data frame containing the survival data.}

\item{time_var}{Name of the time variable in \code{data} (string).}

\item{event_var}{Name of the event variable in \code{data} (string). Expected values are \code{"Dead"} and
\code{"Censored"}.}

\item{var}{Name of the variable to facet by (string).}

\item{nbins}{Number of histogram bins. Defaults to \code{max(data[[time_var]])}.}

\item{ncol}{Number of columns in the facet layout. Defaults to 1.}
}
\value{
A \code{ggplot2} object, returned invisibly. The plot is also
printed as a side effect.
}
\description{
Plots a faceted histogram of times to event (death or censoring), coloured by exit reason.
Useful for visually inspecting the distribution of survival times and censoring
across subgroups.
}
\examples{
\dontrun{
f_t_census(data = my_data, time_var = "time", event_var = "status", var = "group")
}

}
