% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAUS-package.R
\docType{package}
\name{TAUS-package}
\alias{TAUS}
\alias{TAUS-package}
\title{TAUS: Target-Age Unified Survival}
\description{
Computes conditional survival P(T>tau|T>t) and the probability O_tau that a
randomly sampled individual of unknown age will outlive tau. No proportional
hazards or parametric survival form. Pairwise comparison of O_tau via
Beta-KS bootstrap.
}
\references{
TAUS manuscript: methodology and application (this repository and associated
thesis/paper materials). Use \code{citation("TAUS")} for the package citation.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/casasgomezuribarri/Survival}
  \item Report bugs at \url{https://github.com/casasgomezuribarri/Survival/issues}
}

}
\author{
Iván Casas Gomez-Uribarri (maintainer) \email{i.casas-gomez-uribarri.1@research.gla.ac.uk}

Simon A. Babayan, Fredros Okumu, Francesco Baldini, Mauro Pazmiño Betancourth
}
\keyword{internal}
